[tests] moved mem layout parsing tests to gef_memory.py
